backend/requirements.txt
# ============================================================
# Internal Docs Q&A Agent — Backend Dependencies
# ============================================================
# Install: pip install -r requirements.txt
# ============================================================

# ── Web Framework & Server ───────────────────────────────────
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# ── Database (PostgreSQL + ORM) ──────────────────────────────
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ── AI / ML — Embeddings & LLM ──────────────────────────────
openai==1.12.0
tiktoken==0.5.2

# ── Vector Database ──────────────────────────────────────────
chromadb==0.4.22

# ── Document Processing ─────────────────────────────────────
pdfplumber==0.10.4
markdown==3.5.2
beautifulsoup4==4.12.3

# ── Authentication & Security ───────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9

# ── Slack Integration ───────────────────────────────────────
slack-bolt==1.18.1
slack-sdk==3.27.1

# ── Caching ─────────────────────────────────────────────────
redis==5.0.1

# ── HTTP ────────────────────────────────────────────────────
httpx==0.26.0
aiofiles==23.2.1

# ── Utilities ───────────────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.2

# ── Testing ─────────────────────────────────────────────────
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
